% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_attributes.R
\name{age_distribution}
\alias{age_distribution}
\title{population age distribution at time}
\usage{
age_distribution(t)
}
\arguments{
\item{t}{double time}
}
\value{
distribution function
}
\description{
population age distribution at time
}
