% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations_weight_calculator.R
\name{calculate_relations_weight}
\alias{calculate_relations_weight}
\title{calculate_relations_weight}
\usage{
calculate_relations_weight(x, settings)
}
\arguments{
\item{x}{a relations data.frame}

\item{settings}{relations_settings object}
}
\value{
a relations data.frame with the additional column weight
}
\description{
Add a weight for every relation.
}
