% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{compare_and_combine_values}
\alias{compare_and_combine_values}
\title{deal with double values}
\usage{
compare_and_combine_values(x)
}
\arguments{
\item{x}{data.frame column}
}
\description{
deal with double values
}
