% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_analysis.R
\name{count_population_by_living_units_over_time}
\alias{count_population_by_living_units_over_time}
\title{count population by living units per timestep in population}
\usage{
count_population_by_living_units_over_time(humans, time)
}
\arguments{
\item{humans}{tibble population}

\item{time}{vector timesteps}
}
\value{
tibble
}
\description{
count population by living units per timestep in population
}
