% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations_double_connections.R
\name{create_network_comparison_string}
\alias{create_network_comparison_string}
\title{generate relations}
\usage{
create_network_comparison_string(x)
}
\arguments{
\item{x}{data.frame}
}
\description{
generate relations
}
