% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{deal_with_double_connections}
\alias{deal_with_double_connections}
\title{deal with double connections}
\usage{
deal_with_double_connections(x)
}
\arguments{
\item{x}{data.frame}
}
\description{
deal with double connections
}
