% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pop.R
\name{flounder}
\alias{flounder}
\title{Title flounder data}
\usage{
flounder(n, seed = 41, roundage = FALSE)
}
\arguments{
\item{n}{sample size}

\item{seed}{seed, initially 41}

\item{roundage}{either return rounded age or not initally FALSE}
}
\value{
data.frame
}
\description{
Title flounder data
}
\examples{

flounder(1000)

}
