% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_population_relations_attributes.R
\name{friendship_age_distribution}
\alias{friendship_age_distribution}
\title{friends age distribution at time}
\usage{
friendship_age_distribution(t)
}
\arguments{
\item{t}{double time}
}
\value{
distribution function
}
\description{
friends age distribution at time
}
