% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_generator.R
\name{generate_graph}
\alias{generate_graph}
\title{construct_graph}
\usage{
generate_graph(rel)
}
\arguments{
\item{rel}{relations data.frame}
}
\value{
igraph graph object
}
\description{
construct_graph
}
