% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations_generator_horizontal.R
\name{generate_horizontal_relations}
\alias{generate_horizontal_relations}
\title{generate horizontal relations}
\usage{
generate_horizontal_relations(settings)
}
\arguments{
\item{settings}{test}
}
\value{
huup
}
\description{
generate horizontal relations
}
