% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_humans_generator.R
\name{generate_humans}
\alias{generate_humans}
\title{generate_humans}
\usage{
generate_humans(start, stop, n, settings)
}
\arguments{
\item{start}{integer start time}

\item{stop}{integer stop time}

\item{n}{integer amount of humans to generate}

\item{settings}{unit_settings object}
}
\value{
data.frame with population (one row for each individual)
}
\description{
generate_humans
}
