% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pop.R
\name{generate_pop}
\alias{generate_pop}
\title{Generates an imaginary population}
\usage{
generate_pop(n)
}
\arguments{
\item{n}{integer}
}
\value{
data frame
}
\description{
Generates an imaginary population
}
