% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_generator.R
\name{generate_population}
\alias{generate_population}
\title{generate_population}
\usage{
generate_population(settings)
}
\arguments{
\item{settings}{population_settings object}
}
\value{
data.frame with population (one row for each individual)
}
\description{
Generate an initial population based on an populations settings
object and let it live over the course of the timeframe. That includes
death and birth of individual humans. Their units die and rise with
them.
}
