% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations_generator.R
\name{generate_relations}
\alias{generate_relations}
\title{generate relations}
\usage{
generate_relations(settings)
}
\arguments{
\item{settings}{relations_settings object}
}
\value{
huup
}
\description{
Generate horizontal and vertical relations for a population
based on a relations_settings object. Vertical and horizontal
relations are created independently and get merged later.
Double connections get removed. Every relationtype has a specific
weight.
}
