% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_generator.R
\name{generate_unit}
\alias{generate_unit}
\title{generate_unit}
\usage{
generate_unit(settings)
}
\arguments{
\item{settings}{unit_settings object}
}
\value{
data.frame with population (one row for each individual)
}
\description{
Generate an initial population based on an populations settings
object
}
