% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations_generator_vertical.R
\name{generate_vertical_relations}
\alias{generate_vertical_relations}
\title{generate_vertical_relations}
\usage{
generate_vertical_relations(settings)
}
\arguments{
\item{settings}{relations_settings object}
}
\value{
relations data.frame (every row contains one relation)
}
\description{
Generate relations between parents and children.
}
