% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_population_relations_attributes.R
\name{get_attribute}
\alias{get_attribute}
\title{get n attributes according to attribute distribution}
\usage{
get_attribute(n = NA, distribution_function, range)
}
\arguments{
\item{n}{integer amount}

\item{distribution_function}{function probability distribution}

\item{range}{vector value range to query}
}
\value{
vector of attributes
}
\description{
get n attributes according to attribute distribution
}
