% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_population_relations_attributes.R
\name{get_relation_weight}
\alias{get_relation_weight}
\title{get n relation weights according to relation type weight distribution at time}
\usage{
get_relation_weight(t = NA, n = NA, distribution_function)
}
\arguments{
\item{t}{double time}

\item{n}{integer amount}

\item{distribution_function}{function probability distribution}
}
\value{
vector of attributes
}
\description{
get n relation weights according to relation type weight distribution at time
}
