% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideas_settings_class.R
\name{idea_distribution_to_starting_nodes}
\alias{idea_distribution_to_starting_nodes}
\title{idea_distribution_to_starting_nodes}
\usage{
idea_distribution_to_starting_nodes(settings, start_time)
}
\arguments{
\item{settings}{ideas_settings object}

\item{start_time}{moment zero in the model context}
}
\value{
a list of numeric vectors
}
\description{
idea_distribution_to_starting_nodes
}
