% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideas_settings_class.R
\docType{class}
\name{ideas_settings-class}
\alias{ideas_settings-class}
\title{ideas_settings class}
\description{
class to store idea definition values
}
\section{Slots}{

\describe{
\item{\code{population}}{test}

\item{\code{names}}{test}

\item{\code{start_distribution}}{test}

\item{\code{strength}}{test}
}}

