% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideas_settings_class.R
\name{init_ideas_settings}
\alias{init_ideas_settings}
\title{init_ideas_settings}
\usage{
init_ideas_settings(x)
}
\arguments{
\item{x}{populations_grid data.frame}
}
\value{
populations_grid data.frame with additional column 
ideas_settings
}
\description{
Create ideas_settings object for every row in the 
populations_grid data.frame and add it in an additional
column.
}
