% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations_generator.R
\name{init_relations_settings}
\alias{init_relations_settings}
\title{init_relations_settings}
\usage{
init_relations_settings(x)
}
\arguments{
\item{x}{all_model_populations}
}
\value{
huup
}
\description{
init_relations_settings
}
