% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_settings_class.R
\name{init_unit_settings}
\alias{init_unit_settings}
\title{init_unit_settings}
\usage{
init_unit_settings(x)
}
\arguments{
\item{x}{populations_settings object}
}
\value{
populations_grid data.frame with additional column 
population_settings
}
\description{
init_unit_settings
}
