% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pop.R
\name{mackerel}
\alias{mackerel}
\title{Generate mackerel data}
\usage{
mackerel(n, seed = 41, roundage = FALSE)
}
\arguments{
\item{n}{sample size}

\item{seed}{seed, initally 41}

\item{roundage}{either return rounded age or not}
}
\value{
data.frame
}
\description{
Generate mackerel data
}
\examples{
mackerel(1000,seed=1, roundage=TRUE)

}
