% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations_modifier_cross_unit.R
\name{modify_relations_cross_unit}
\alias{modify_relations_cross_unit}
\title{modify_relations_cross_unit}
\usage{
modify_relations_cross_unit(relations, cross_unit_proportion,
  unit_interaction_matrix, population)
}
\arguments{
\item{relations}{relations data.frame}

\item{cross_unit_proportion}{cross_unit_proportion}

\item{unit_interaction_matrix}{unit_interaction_matrix}

\item{population}{population data.frame}
}
\value{
modified relations data.frame
}
\description{
Replace existing relations within units by relations across units.
}
