% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph_properties.R
\name{plot_population_development}
\alias{plot_population_development}
\title{plot_population_development}
\usage{
plot_population_development(pop, timeframe = c())
}
\arguments{
\item{pop}{population data.frame}

\item{timeframe}{timeframe}
}
\value{
ggplot2 plot object
}
\description{
plot_population_development
}
