% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep_grid.R
\name{plot_prep_grid}
\alias{plot_prep_grid}
\title{Plot grid attributes}
\usage{
plot_prep_grid(x, method)
}
\arguments{
\item{x}{grid}

\item{method}{test}
}
\value{
plot
}
\description{
Plot grid attributes
}
