% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph_properties.R
\name{plot_relations_development}
\alias{plot_relations_development}
\title{plot_relations_development}
\usage{
plot_relations_development(pop, rel, time = c())
}
\arguments{
\item{pop}{population data.frame}

\item{rel}{relations data.frame}

\item{time}{time}
}
\value{
ggplot2 plot object
}
\description{
plot_relations_development
}
