% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_model_preparation.R
\name{popgenerator}
\alias{popgenerator}
\title{popgenerator}
\usage{
popgenerator(timeframe, unit_amount, unit_names, unit_size_functions,
  age_distribution_function, age_range, amounts_friends,
  unit_interaction_matrix, cross_unit_proportion_child_of,
  cross_unit_proportion_friend, weight_child_of, weight_friend, dir_path,
  only_idea_proportions = TRUE)
}
\arguments{
\item{dir_path}{directory path where to store the output files}

\item{only_idea_proportions}{should only the final result file be stored permanently}

\item{x}{models_grid data.frame}

\item{cores}{how many parallel threads should be used}
}
\value{
modified models_grid with additional columns
}
\description{
popgenerator
}
