% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_settings_class.R
\docType{class}
\name{population_settings-class}
\alias{population_settings-class}
\title{population_settings class}
\description{
class to store values that should be passed to the population generator
}
\section{Slots}{

\describe{
\item{\code{time}}{test}

\item{\code{population_size_function}}{test}

\item{\code{unit_amount_function}}{test}

\item{\code{age_distribution_function}}{test}

\item{\code{age_range}}{test}

\item{\code{sex_distribution_function}}{test}

\item{\code{sex_range}}{test}

\item{\code{unit_distribution_function}}{test}

\item{\code{start_unit_vector}}{test}
}}

