% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_attributes.R
\name{population_size}
\alias{population_size}
\title{population size at time}
\usage{
population_size(t)
}
\arguments{
\item{t}{double time}
}
\value{
integer population size at t
}
\description{
population size at time
}
