% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_model_preparation.R
\name{prepare_and_export_landscapes}
\alias{prepare_and_export_landscapes}
\title{prepare_and_export_landscapes}
\usage{
prepare_and_export_landscapes(x, dir_path,
  cores = parallel::detectCores())
}
\arguments{
\item{x}{models_grid data.frame}

\item{dir_path}{directory path where to store the output files}

\item{cores}{how many parallel threads should be used}
}
\value{
modified models_grid with additional columns
}
\description{
prepare_and_export_landscapes
}
