% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_model_preparation.R
\name{prepare_pops_rels_ideas}
\alias{prepare_pops_rels_ideas}
\title{prepare_pops_rels_ideas}
\usage{
prepare_pops_rels_ideas(x)
}
\arguments{
\item{x}{models_grid data.frame}
}
\value{
modified models_grid with additional columns
}
\description{
prepare_pops_rels_ideas
}
