% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomage.R
\name{randomage}
\alias{randomage}
\title{Generate random ages}
\usage{
randomage(n, mean, sd, roundage = FALSE)
}
\arguments{
\item{n}{integer, number of age generated}

\item{mean}{positive numeric}

\item{sd}{positive numeric}

\item{roundage}{round age or allow dicimals. The function will return with ceiling.}
}
\value{
vector
}
\description{
Generate random ages
}
\examples{
randomage(n=100, mean=2, sd=1, roundage=TRUE)
randomage(n=10, mean=3, sd=1.5)
}
