% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations_settings_class.R
\docType{class}
\name{relations_settings-class}
\alias{relations_settings-class}
\title{relations_settings class}
\description{
class to store values that should be passed to the relations generator
}
\section{Slots}{

\describe{
\item{\code{population}}{test}

\item{\code{amount_friends}}{test}

\item{\code{cross_unit_proportion_child_of}}{test}

\item{\code{cross_unit_proportion_friend}}{test}

\item{\code{weight_child_of}}{test}

\item{\code{weight_friend}}{test}
}}

