% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helper_functions.R
\name{resample}
\alias{resample}
\title{resample}
\usage{
resample(x, ...)
}
\arguments{
\item{x}{a vector of one or more elements from which to choose}

\item{...}{arguments passed to sample.int()}
}
\value{
vector of attributes
}
\description{
better base::sample.int() implementation
}
