% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluesless_run.R
\name{run_gluesless}
\alias{run_gluesless}
\title{run_gluesless}
\usage{
run_gluesless(app_path, input_file_dir, output_file_path, models_to_run)
}
\arguments{
\item{app_path}{test}

\item{input_file_dir}{test}

\item{output_file_path}{test}

\item{models_to_run}{test}
}
\value{
a result list
}
\description{
run_gluesless
}
