% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_model_preparation.R
\name{run_simulation}
\alias{run_simulation}
\title{run_simulation}
\usage{
run_simulation(x, dir_path, only_idea_proportions = TRUE,
  cores = parallel::detectCores())
}
\arguments{
\item{x}{models_grid data.frame}

\item{dir_path}{directory path where to store the output files}

\item{only_idea_proportions}{should only the final result file be stored permanently}

\item{cores}{how many parallel threads should be used}
}
\value{
modified models_grid with additional columns
}
\description{
run_simulation
}
