% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pop.R
\name{sardine}
\alias{sardine}
\title{Generate sardine data}
\usage{
sardine(n, seed = 41, roundage = FALSE)
}
\arguments{
\item{n}{sample size}

\item{seed}{seed, initially 41}

\item{roundage}{either retrun rounded age or not}
}
\value{
data frame
}
\description{
Generate sardine data
}
\examples{
sardine(1000, seed=1, roundage=FALSE)

}
