% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_attributes.R
\name{sex_distribution}
\alias{sex_distribution}
\title{population sex distribution at time}
\usage{
sex_distribution(t)
}
\arguments{
\item{t}{double time}
}
\value{
distribution function
}
\description{
population sex distribution at time
}
