% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_growth_simulator.R
\name{simulate_growth}
\alias{simulate_growth}
\title{Simulate growth of a population.}
\usage{
simulate_growth(humans, settings)
}
\arguments{
\item{humans}{tibble input population}

\item{settings}{population_settings object}
}
\value{
data.frame with population (one row for each individual)
}
\description{
Simulate growth of a population.
}
