% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_attributes.R
\name{unit_amount}
\alias{unit_amount}
\title{unit amount at time}
\usage{
unit_amount(t)
}
\arguments{
\item{t}{double time}
}
\value{
integer amount of units at t
}
\description{
unit amount at time
}
