% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_settings_class.R
\docType{class}
\name{unit_settings-class}
\alias{unit_settings-class}
\title{population_settings class}
\description{
class to store values that should be passed to the population generator
}
\section{Slots}{

\describe{
\item{\code{time}}{test}

\item{\code{unit_name}}{test}

\item{\code{unit_size_function}}{test}

\item{\code{age_distribution_function}}{test}

\item{\code{age_range}}{test}
}}

