% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_population_relations_attributes.R
\name{weight_child_of_distribution}
\alias{weight_child_of_distribution}
\title{weight child_of distribution at time}
\usage{
weight_child_of_distribution(t)
}
\arguments{
\item{t}{double time}
}
\value{
distribution function
}
\description{
weight child_of distribution at time
}
