% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_functions.R
\name{write_all_models_to_files}
\alias{write_all_models_to_files}
\title{write_all_models_to_files}
\usage{
write_all_models_to_files(populations, relations, ideas_settings,
  timeframe, model_id, dir_path)
}
\arguments{
\item{populations}{list of populations}

\item{relations}{list of relations}

\item{ideas_settings}{list of ideas_settings}

\item{timeframe}{list of timeframes}

\item{model_id}{vector of model ids}

\item{dir_path}{directory path where to store the output files}
}
\value{
called for side effect writing to file system
}
\description{
write_all_models_to_files
}
