% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_functions.R
\name{write_idea_proportions}
\alias{write_idea_proportions}
\title{write_idea_proportions}
\usage{
write_idea_proportions(model_id, idea_proportions, dir_path)
}
\arguments{
\item{model_id}{model_id}

\item{idea_proportions}{idea_proportions list}

\item{dir_path}{output dir path}
}
\value{
TRUE, called for the side effect of writing to the file system
}
\description{
write_idea_proportions
}
