% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_and_export_functions.R
\name{write_pajek}
\alias{write_pajek}
\title{write_pajek}
\usage{
write_pajek(graph, pop_df, path)
}
\arguments{
\item{graph}{igraph object}

\item{pop_df}{population data.frame}

\item{path}{output file path}
}
\value{
TRUE, called for the side effect of writing to the file system
}
\description{
write_pajek
}
