% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_functions.R
\name{write_pajek_for_snap}
\alias{write_pajek_for_snap}
\title{write_pajek_for_snap}
\usage{
write_pajek_for_snap(rel, pop, path, simple_version_path = NULL)
}
\arguments{
\item{rel}{relations data.frame}

\item{pop}{population data.frame}

\item{path}{output file path}

\item{simple_version_path}{output file path for reduced outfile file version}
}
\value{
TRUE, called for the side effect of writing to the file system
}
\description{
write_pajek_for_snap
}
