% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_analysis.R
\name{calculate_all_idea_proportions_over_time}
\alias{calculate_all_idea_proportions_over_time}
\title{calculate_all_idea_proportions_over_time}
\usage{
calculate_all_idea_proportions_over_time(model_id, populations, timeframe,
  model_group, simulation_results, by_unit = FALSE)
}
\arguments{
\item{model_id}{test}

\item{populations}{test}

\item{timeframe}{test}

\item{model_group}{test}

\item{simulation_results}{test}

\item{by_unit}{test}
}
\value{
test
}
\description{
calculate_all_idea_proportions_over_time
}
