% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_analysis.R
\name{calculate_idea_proportions_over_time}
\alias{calculate_idea_proportions_over_time}
\title{calculate_idea_proportions_over_time}
\usage{
calculate_idea_proportions_over_time(id, x)
}
\arguments{
\item{id}{test}

\item{x}{test}
}
\value{
test
}
\description{
calculate_idea_proportions_over_time
}
