% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations_weight_calculator.R
\name{calculate_relations_weight}
\alias{calculate_relations_weight}
\title{calculate relations weight}
\usage{
calculate_relations_weight(x, settings)
}
\arguments{
\item{x}{a relations table}

\item{settings}{relations_settings object}
}
\value{
huup
}
\description{
calculate relations weight
}
