% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_generator.R
\name{generate_all_populations}
\alias{generate_all_populations}
\title{generate_all_populations}
\usage{
generate_all_populations(x)
}
\arguments{
\item{x}{populations_grid data.frame}
}
\value{
populations_grid data.frame with additional column
populations
}
\description{
Generate populations in populations_grid data.frame based on
population_settings objects in column population_settings.
}
