% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations_generator.R
\name{generate_all_relations}
\alias{generate_all_relations}
\title{generate_all_relations}
\usage{
generate_all_relations(x)
}
\arguments{
\item{x}{populations_grid data.frame}
}
\value{
populations_grid data.frame with additional column
relations
}
\description{
Generate relations in a populations_grid data.frame based on
relations_settings objects in column relations_settings.
}
