% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations_generator_horizontal.R
\name{generate_horizontal_relations}
\alias{generate_horizontal_relations}
\title{generate_horizontal_relations}
\usage{
generate_horizontal_relations(settings)
}
\arguments{
\item{settings}{relations_settings object}
}
\value{
relations data.frame (every row contains one relation)
}
\description{
Generate relations between random individuals.
}
