% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pop.R
\name{generate_pop}
\alias{generate_pop}
\title{Generates an imaginary population}
\usage{
generate_pop(n, p, roundage = FALSE)
}
\arguments{
\item{n}{Number of samples to be generated}

\item{p}{parameters for growth (p[1],p[2],p[3]) and mean_age p[4] and its sd p[5]}

\item{roundage}{Rounds the age to integer if 'TRUE'. Default is set as 'FALSE' which returns ages with dicimals.}
}
\value{
data frame
}
\description{
Generates an imaginary population
}
\details{
Fishtype can be selected from "sardine", "flounder", "mackerel", and "tuna". When type = "manual", then VBGF parameters need to be input manually as a vector of c(Linf, K, t0).
"sardine" consists of short lived organism with small body size. This species exhibit indeterminate growth with inflexion point of growth around 2 years old.
}
\examples{

generate_pop(1000,c(524,  0.19,  -1.61, 4, 2),TRUE)

}
\references{
Manabe, A., et al. (2018). A novel ontogenetic function incorporating energy allocation. PLoS ONE. e######.
}
