% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_generator.R
\name{generate_population}
\alias{generate_population}
\title{generate population}
\usage{
generate_population(settings)
}
\arguments{
\item{settings}{test}
}
\value{
huup
}
\description{
generate population
}
