% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations_generator.R
\name{generate_relations}
\alias{generate_relations}
\title{generate relations}
\usage{
generate_relations(settings)
}
\arguments{
\item{settings}{test}
}
\value{
huup
}
\description{
generate relations
}
