% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations_generator_vertical.R
\name{generate_vertical_relations}
\alias{generate_vertical_relations}
\title{generate vertical relations}
\usage{
generate_vertical_relations(settings)
}
\arguments{
\item{settings}{test}
}
\value{
huup
}
\description{
generate vertical relations
}
