% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_attributes.R
\name{get_attribute}
\alias{get_attribute}
\title{get n attributes according to attribute distribution at time}
\usage{
get_attribute(t = NA, n = NA, distribution_function, range)
}
\arguments{
\item{t}{double time}

\item{n}{integer amount}

\item{distribution_function}{function probability distribution}

\item{range}{vector value range to query}
}
\value{
vector of attributes
}
\description{
get n attributes according to attribute distribution at time
}
