% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_generator.R
\name{init_population_settings}
\alias{init_population_settings}
\title{init_population_settings}
\usage{
init_population_settings(x)
}
\arguments{
\item{x}{all_model_populations}
}
\value{
huup
}
\description{
init_population_settings
}
