% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_settings_class.R
\name{init_population_settings}
\alias{init_population_settings}
\title{init_population_settings}
\usage{
init_population_settings(time, unit_names, unit_size_functions,
  age_distribution_function, age_range)
}
\arguments{
\item{time}{test}

\item{unit_names}{test}

\item{unit_size_functions}{test}

\item{age_distribution_function}{test}

\item{age_range}{test}

\item{unit_amount}{test}
}
\value{
populations_grid data.frame with additional column 
population_settings
}
\description{
Create population_settings object for every row in the 
populations_grid data.frame and add it in an additional
column.
}
