% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations_settings_class.R
\name{init_relations_settings}
\alias{init_relations_settings}
\title{init_relations_settings}
\usage{
init_relations_settings(population, amount_friends,
  unit_interaction_matrix, cross_unit_proportion_child_of,
  cross_unit_proportion_friend, weight_child_of, weight_friend)
}
\arguments{
\item{population}{test}

\item{amount_friends}{test}

\item{unit_interaction_matrix}{test}

\item{cross_unit_proportion_child_of}{test}

\item{cross_unit_proportion_friend}{test}

\item{weight_child_of}{test}

\item{weight_friend}{test}
}
\value{
populations_grid data.frame with additional column 
relations_settings
}
\description{
Create relations_settings object for every row in the 
populations_grid data.frame and add it in an additional
column.
}
