% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations_settings_class.R
\name{init_relations_settings}
\alias{init_relations_settings}
\title{init_relations_settings}
\usage{
init_relations_settings(x, populations)
}
\arguments{
\item{x}{populations_grid data.frame}

\item{populations}{populations list}
}
\value{
populations_grid data.frame with additional column 
relations_settings
}
\description{
Create relations_settings object for every row in the 
populations_grid data.frame and add it in an additional
column.
}
