% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations_modifier_cross_unit.R
\name{modify_relations_cross_unit}
\alias{modify_relations_cross_unit}
\title{modify_relations_cross_unit}
\usage{
modify_relations_cross_unit(relations, settings)
}
\arguments{
\item{relations}{relations data.frame}

\item{settings}{relations_settings object}
}
\value{
modified relations data.frame
}
\description{
Replace existing relations within units by relations across units.
}
