% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations_settings_class.R
\docType{class}
\name{relations_settings-class}
\alias{relations_settings-class}
\title{relations_settings class}
\description{
class to store values that should be passed to the relations generator
}
\section{Slots}{

\describe{
\item{\code{population}}{test}

\item{\code{monogamy_probability}}{test}

\item{\code{start_fertility_age}}{test}

\item{\code{stop_fertility_age}}{test}

\item{\code{same_unit_as_child_probability}}{test}

\item{\code{same_unit_as_partner_probability}}{test}
}}

