% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_attributes.R
\name{unit_distribution}
\alias{unit_distribution}
\title{population unit distribution at time}
\usage{
unit_distribution(t)
}
\arguments{
\item{t}{double time}
}
\value{
distribution function
}
\description{
population unit distribution at time
}
