% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_functions.R
\name{write_all_models_to_files}
\alias{write_all_models_to_files}
\title{write_all_models_to_files}
\usage{
write_all_models_to_files(x, dir_path)
}
\arguments{
\item{x}{models_grid data.frame}

\item{dir_path}{directory path where to store the files}
}
\value{
called for side effect writing to file system
}
\description{
write_all_models_to_files
}
